Fix retrieving review message.
In case of NOT_BUILT status (for ex. when a build was removed from
the queue before it has been started) we get null in both labels:
Verify and Code-Review.
In this case we remove these labels from result command. After that
assuming review message is located between "--message '" and "'  --"
fails due to we got extra spaces after labels has been removed.
As a result, an option following the message is parsed as a part
of the message.